copy from trunk to get cs
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/gus/CBIL/branches/4.0@18728 4235af6a-31f9-0310-9ad3-ecf6348f2534

+++ Util/lib/perl/Sra.pm
@@ -6,7 +6,7 @@ use Data::Dumper;
 
 require Exporter;
 @ISA = qw(Exporter);
-@EXPORT = qw(getRunIdsFromSraSampleIds getFastqForSraRunId getFastqForSampleIds);
+@EXPORT = qw(getRunIdsFromSraSampleIds getFastqForSraRunId getFastqForSampleIds getCsForSampleIds);
 
 
 sub getFastqForSampleIds {
@@ -80,6 +80,76 @@ sub getFastqForSampleIds {
   rename("tmpReads_1.fastq","$fileoutone") if (-e "tmpReads_1.fastq");
   rename("tmpReads_2.fastq","$fileouttwo") if (-e "tmpReads_2.fastq");
 }
+
+sub getCsForSampleIds {
+  my($sids,$fileoutone,$fileouttwo,$dontdownload) = @_;
+  $fileoutone = $fileoutone ? $fileoutone : "reads_1.csfasta";
+  $fileouttwo = $fileouttwo ? $fileouttwo : "reads_2.csfasta";
+  my @rids;
+  foreach my $sid (@{$sids}){
+    push(@rids,&getRunIdsFromSraSampleId($sid));
+  }
+  my @out;
+  my $readCount = 0;
+  my $fid = $rids[0]->[1];
+  my $singleEnd = 0;
+  my $doubleEnd = 0;
+  my %tsid;
+  my %done;
+  foreach my $a (@rids){
+    $tsid{$a->[0]} = 1;
+    if($done{$a->[1]}){
+      print STDERR "ERROR: already retrieved '$a->[1]' .. skipping\n";
+      next;
+    }
+    push(@out,"$a->[0]:$a->[1]:$a->[2]:".($a->[2] ? $a->[3] / $a->[2] : 'undef').":$a->[4]");
+    $readCount += $a->[2];
+    my $id = $a->[1];
+    $done{$id} = 1;
+    next if $dontdownload;
+    &getCsForSraRunId($id);
+    ##if single end will have two files labeled _F3.csfasta and _F3_QV.qual .. otherwise four labeled _F3.csfasta and _F3_QV.qual, _R3.csfasta and _R3_QV.qual
+    my $foundFile = 0;
+    if(-e "$id\_F3.csfasta" && -e "$id\_F3_QV.qual"){
+      $foundFile++;
+      if($fid ne $id){
+        system("cat $id\_F3.csfasta >> tmpReads_1.csfasta");
+        system("cat $id\_F3_QV.qual >> tmpReads_1.csfasta.qual");
+        unlink("$id\_F3.csfasta");
+        unlink("$id\_F3.csfasta");
+      }else{
+        rename("$id\_F3.csfasta","tmpReads_1.csfasta");
+        rename("$id\_F3_QV.qual","tmpReads_1.csfasta.qual");
+      }
+    }
+    if(-e "$id\_R3.csfasta" && -e "$id\_R3_QV.qual"){
+      $doubleEnd = 1;
+      die "ERROR: this sample '$a->[0]' contains both single and double end reads\n" if $singleEnd && $doubleEnd;
+      $foundFile++;
+      if($fid ne $id){
+        system("cat $id\_R3.csfasta >> tmpReads_2.csfasta");
+        system("cat $id\_R3_QV.qual >> tmpReads_2.csfasta.qual");
+        unlink("$id\_R3.csfasta");
+        unlink("$id\_R3.csfasta");
+      }else{
+        rename("$id\_R3.csfasta","tmpReads_2.csfasta");
+        rename("$id\_R3_QV.qual","tmpReads_2.csfasta.qual");
+      }
+    }else{
+      ##is single end only
+      $singleEnd = 1;
+      die "ERROR: this sample '$a->[0]' contains both single and double end reads\n" if $singleEnd && $doubleEnd;
+    }
+  } 
+  print "input: (",join(", ",@{$sids}),") ", scalar(keys%tsid), " samples, ", scalar(@rids) , " runs, $readCount spots: " , (scalar(@rids) == 0 ? "ERROR: unable to retrieve runIds\n" : "(",join(", ",@out),")\n");
+  ##now mv the files to a final filename ...
+
+  rename("tmpReads_1.csfasta","$fileoutone") if (-e "tmpReads_1.csfasta");
+  rename("tmpReads_2.csfasta","$fileouttwo") if (-e "tmpReads_2.csfasta");
+  rename("tmpReads_1.csfasta.qual","$fileoutone.qual") if (-e "tmpReads_1.csfasta.qual");
+  rename("tmpReads_2.csfasta.qual","$fileouttwo.qual") if (-e "tmpReads_2.csfasta.qual");
+}
+
 
 sub getRunIdsFromSraSampleId { 
  my ($sid) = @_; 
@@ -175,6 +245,23 @@ sub getFastqForSraRunId {
   print STDERR "extracting fastq file(s)...";
   system("fastq-dump --split-files ./$file");
   my @files = glob("$runId*.fastq");
+  print STDERR "DONE: ".scalar(@files)." files (".join(", ",@files).")\n";
+  unlink($file);
+}
+
+sub getCsForSraRunId {
+  my($runId,$pe) = @_;
+  my $file = "$runId.sra";
+  unlink("$runId.sra") if -e "$runId.sra";
+  my $cmd = "wget http://ftp-private.ncbi.nlm.nih.gov/sra/sra-instant/reads/ByRun/sra/".substr($runId,0,3)."/".substr($runId,0,6)."/$runId/$file";
+  print STDERR "retrieving $runId with $cmd\n";
+  system($cmd);
+  if($?){
+    die "ERROR ($?): Unable to fetch sra file for $runId\n";
+  }
+  print STDERR "extracting fastq file(s)...";
+  system("abi-dump ./$file");
+  my @files = glob("$runId*.*");
   print STDERR "DONE: ".scalar(@files)." files (".join(", ",@files).")\n";
   unlink($file);
 }
